統合テストのデプロイコードと、実際のデプロイコードを合わせる
